18th code

+++ 18-lineintersection-Python/lineintersection.py
@@ -12,7 +12,8 @@ def lineintersection(m1, b1, m2, b2):
 		return None
 	elif (m2 % m1 == 0):
 		return None
-	return b2-b1/m1-m2
+	# print(b2,b1,m1,m2)
+	return int((b2-b1)/(m1-m2))
 	# pass
 
 if __name__ == "__main__":
